Criar o widget Home e refatoração
