--- lib/pages/dashBoard/views/dashBoardBody.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-import '../../dashboardAulas/dashboard_view.dart';
-
-class DashBoardBody extends StatelessWidget {
-  const DashBoardBody({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(children: [
-      Expanded(child: Text("")),
-    ]);
-  }
-}
