--- lib/pages/dashBoard/views/dashBoardMainPage.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-class DashBoardMainPage extends StatelessWidget {
-  const DashBoardMainPage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.blueAccent,
-      child: Center(child: Text('weeee')),
-    );
-  }
-}
