+++ lib/pages/dashboard_classes/dashboard_classes_view.dart
@@ -1,0 +1,31 @@
+import 'package:conquistandoomundo/pages/dashboard_classes/views/header.dart';
+import 'package:conquistandoomundo/pages/dashboard_classes/views/youtube_player.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class DashboardClasses extends StatefulWidget {
+  List<String> ids = [];
+  DashboardClasses(this.ids, {Key? key}) : super(key: key);
+
+  @override
+  State<DashboardClasses> createState() => _DashboardClassesState();
+}
+
+class _DashboardClassesState extends State<DashboardClasses> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: Header(),
+      body: MaterialApp(
+        title: 'Conquistando o mundo',
+        theme: ThemeData(
+          brightness: Brightness.dark,
+          primarySwatch: Colors.deepPurple,
+          scaffoldBackgroundColor: Color.fromARGB(255, 11, 20, 35),
+        ),
+        debugShowCheckedModeBanner: false,
+        home: YoutubePlayer(widget.ids),
+      ),
+    );
+  }
+}
